File(fix): seek support uint64_t and not size_t

+++ main/system/File.hpp
@@ -14,7 +14,7 @@ public:
     std::vector<uint8_t> read_all_bytes();
     std::vector<uint8_t> read_bytes(const size_t size);
     void write(const std::string& txt);
-    void seek(size_t position);
+    void seek(uint64_t position);
     uint64_t size();
 
 private:
